+++ src/layouts/PublicLayout.js
@@ -1,5 +1,8 @@
 import React, { Component } from 'react';
 
+// we need to finish it. actually it's a wrapper for pages.
+// like HomeView & HomeRecipeView has same parts, that can be put into container
+
 const PublicLayout = () => {
 
 }
